Add files via upload
This project explores the FIFA 22 player dataset, a comprehensive collection of data extracted from the popular FIFA 22 soccer simulation video game. This dataset was chosen because it reflects in detail the attributes and skills of real-world soccer players, giving us an idea of what makes a successful soccer player. The data comes from the game's vast database and includes a variety of metrics such as player ratings, physical characteristics, and in-game performance statistics. Our goal is to analyze these attributes to gain insight into the factors that influence player strength and market value in both the virtual and real soccer worlds.